# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "To-Do API",
    "description": "API REST para gerenciamento de to-dos com compartilhamento entre usuários. Construída com Next.js + TypeScript + Supabase.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desenvolvimento"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido no login"
      }
    },
...
  },
  "paths": {
    "/api/auth/signup": {
      "post": {
        "summary": "Registra um novo usuário",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "usuario@exemplo.com" },
                  "password": { "type": "string", "minLength": 6, "example": "senha123" },
                  "name": { "type": "string", "example": "João Silva" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário criado com sucesso" },
          "400": { "description": "Dados inválidos" },
          "409": { "description": "Email já está em uso" }
        }
      }
    },
    "/api/auth/signin": {
      "post": {
        "summary": "Realiza login do usuário",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login realizado com sucesso" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
    "/api/auth/signout": {
      "post": {
        "summary": "Realiza logout do usuário",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Logout realizado com sucesso" },
          "401": { "description": "Não autenticado" }
        }
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "summary": "Renova o token de acesso",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refresh_token"],
                "properties": {
                  "refresh_token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token renovado com sucesso" },
          "401": { "description": "Token inválido ou expirado" }
        }
      }
    },
...
    "/api/todos/{id}": {
      "get": {
        "summary": "Busca um to-do específico",
        "tags": ["Todos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Dados do to-do" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Sem permissão" },
          "404": { "description": "To-do não encontrado" }
        }
      },
      "put": {
        "summary": "Atualiza um to-do",
        "tags": ["Todos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "completed": { "type": "boolean" },
                  "priority": { "type": "string", "enum": ["low", "medium", "high"] },
                  "due_date": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "To-do atualizado" },
          "400": { "description": "Dados inválidos" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Sem permissão" },
          "404": { "description": "To-do não encontrado" }
        }
      },
      "delete": {
        "summary": "Deleta um to-do",
        "tags": ["Todos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "To-do deletado" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Sem permissão (apenas o dono pode deletar)" },
          "404": { "description": "To-do não encontrado" }
        }
      }
    },
...
  }
}